+++ shared-components/src/main/scala/org/fiume/sketch/shared/app/ErrorCode.scala
@@ -1,0 +1,4 @@
+package org.fiume.sketch.shared.app
+
+enum ErrorCode:
+  case InvalidCredentials
