+++ shared-components/src/test/resources/contract/shared/app/http/error.info.json
@@ -1,5 +1,5 @@
 {
-  "code": "InvalidCredentials",
+  "code": "INVALID_CREDENTIALS",
   "details": null,
   "message": "The username or password provided is incorrect.",
   "timestamp": null
